% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ari}
\alias{ari}
\title{Adjusted Rand Index}
\usage{
ari(c1, c2)
}
\arguments{
\item{c1}{a vector containing the labels of the first classification. Must be a vector of characters, integers, numerics, or a factor, but not a list.}

\item{c2}{a vector containing the labels of the second classification.}
}
\value{
a scalar with the adjusted rand index.
}
\description{
A function to compute the adjusted rand index between two classifications
}
\examples{
## calculate Adjusted Rand Index on two  sets of labels
data(sceiad_subset_data)
ari(sceiad_subset_data$CellType_predict, sceiad_subset_data$cluster)
}
