% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_sumZscore}
\alias{calc_sumZscore}
\title{Calc_sumZscore}
\usage{
calc_sumZscore(df, batch_key)
}
\arguments{
\item{df}{a dataframe generated by run_all_metrics}

\item{batch_key}{name of batch column in metadata used when generating run_all_metrics}
}
\description{
Aggregate multiple integration metrics across multiple integration runs
}
