% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countPairs}
\alias{countPairs}
\title{Count pairs in vectors}
\usage{
countPairs(classi1, classi2, order)
}
\arguments{
\item{classi1}{first vector to compare}

\item{classi2}{second vector to compare}

\item{order}{proper order of vectors}
}
\description{
Count pairs in vectors
}
