% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_all_metrics}
\alias{run_all_metrics}
\title{Running All Metrics}
\usage{
run_all_metrics(
  reduction,
  metadata,
  batch_key,
  label1_key,
  label2_key,
  run_name = NULL,
  sil_width_prop = 1,
  sil_width_group_key = NULL,
  quietly = F
)
}
\arguments{
\item{reduction}{A matrix of reduced dimensions}

\item{metadata}{A data.frame containing information like batch, cell type, etc}

\item{batch_key}{Name of column in metadata corresponding to batch}

\item{label1_key}{Name of column in metadata corresponding to primary cell label, eg Cell type}

\item{label2_key}{Name of column in metadata corresponding to secondary cell label, eg cluster identity}

\item{run_name}{(optional) name to refer to dataset}

\item{sil_width_prop}{(optional) proportion of data to use for \code{\link{silhouette_width}}}

\item{sil_width_group_key}{(optional) which column in metadata to use for stratified sampling of data}

\item{quietly}{(optional) if TRUE dont print anything}
}
\value{
A one row \code{data.frame} of calculated metrics
}
\description{
Running All Metrics
}
