% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{sceiad_subset_data}
\alias{sceiad_subset_data}
\title{Example scRNA-seq data from the single cell eye in a disk(sceiad)
the original data set this was pulled from can be found at this link
'https://hpc.nih.gov/~mcgaugheyd/scEiaD/colab/scEiaD_all_anndata_mini_ref.h5ad'}
\format{
An object of class \code{"data.frame"}
}
\source{
\<"https://hpc.nih.gov/~mcgaugheyd/scEiaD/colab/scEiaD_all_anndata_mini_ref.h5ad"?>
}
\usage{
data(sceiad_subset_data)
}
\description{
Example scRNA-seq data from the single cell eye in a disk(sceiad)
the original data set this was pulled from can be found at this link
'https://hpc.nih.gov/~mcgaugheyd/scEiaD/colab/scEiaD_all_anndata_mini_ref.h5ad'
}
\examples{
data(sceiad_subset_data)
head(sceiad_subset_data)
}
\keyword{datasets}
