% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPOP.R
\docType{package}
\name{scPOP}
\alias{scPOP}
\title{scPOP: Metrics for Benchmarking scRNA-Seq Batch Correction}
\description{
Evaluate using batch effect correction for scRNA-seq using multiple established methods, including the Adjusted Rand Index,
Normalized Mutual Information, Local Inverse Simpson Index, and Silhouette Width. We also included metrics for #' aggregating and weighing multiple metrics together.
}
